refactors code separating responsibilities
