+++ web/index.html
@@ -318,6 +318,7 @@
     <script type="module" src="assets/js/utils/localStorage.js"></script>
     <script type="module" src="assets/js/utils/render-functions.js"></script>
     <script type="module" src="assets/js/main.js"></script>
+    <script type="module" src="assets/js/share.js"></script>
     <script
       type="module"
       src="assets/js/components/modals/playground-mode.js"
